* Fixed next configuration calculation in case output validation failed * Fixed compiler errors and warnings

+++ examples/gemm_batch/gemm_batch.cpp
@@ -95,7 +95,7 @@ public:
         ktt::DimensionVector globalSize(1, 1, 1);
         ktt::DimensionVector localSize(1, 1, 1);
         std::vector<ktt::ParameterPair> parameterValues = getCurrentConfiguration();
-        int gran = getParameterValue("GRANULARITY", parameterValues);
+        size_t gran = getParameterValue("GRANULARITY", parameterValues);
         if (gran == 1) {
 #if USE_CUDA == 0
             globalSize.setSizeX(batch);
@@ -105,7 +105,7 @@ public:
             localSize.setSizeX(getParameterValue("GROUP_SIZE_X", parameterValues));
         }
         if (gran == 2) {
-            int y = getParameterValue("MGCG_GROUP_SIZE_Y", parameterValues);
+            size_t y = getParameterValue("MGCG_GROUP_SIZE_Y", parameterValues);
 #if USE_CUDA == 0
             globalSize.setSizeX(batch*c / y);
             globalSize.setSizeY(y);
@@ -116,7 +116,7 @@ public:
             localSize.setSizeY(y);
         }
         if (gran == 3) {
-            int y = getParameterValue("MGCG_GROUP_SIZE_Y", parameterValues);
+            size_t y = getParameterValue("MGCG_GROUP_SIZE_Y", parameterValues);
 #if USE_CUDA == 0
             globalSize.setSizeX(batch*c);
             globalSize.setSizeY(y);
@@ -173,11 +173,11 @@ int main(int argc, char** argv)
     std::vector<REAL> dst(c*b*batch, 0.0f);
 
     // fill with random data
-    for (int i = 0; i < a*b*batch; i++)
+    for (size_t i = 0; i < a*b*batch; i++)
     {
         srcA[i] = 10.0f*((REAL)rand()) / ((REAL) RAND_MAX);
     }
-    for (int i = 0; i < c*a*batch; i++)
+    for (size_t i = 0; i < c*a*batch; i++)
     {
         srcB[i] = 10.0f*((REAL)rand()) / ((REAL) RAND_MAX);
     }
